@@ -10,7 +10,7 @@
     "webpack-bundle-analyzer": "webpack-bundle-analyzer"
   },
   "browserslist": [
-      "defaults"
+    "defaults"
   ],
   "devDependencies": {
     "autoprefixer": "^9.6.0",
@@ -18,7 +18,7 @@
     "cssnano": "^4.1.4",
     "express": "^4.14.1",
     "express-http-proxy": "^0.11.0",
-    "gulp": "^4.0.0",
+    "gulp": "^4.0.2",
     "gulp-clean-css": "^4.0.0",
     "gulp-livereload": "^4.0.1",
     "gulp-postcss": "^8.0.0",
@@ -31,56 +31,55 @@
     "postcss-inline-svg": "^3.0.0",
     "pump": "^1.0.2",
     "source-map-loader": "^0.2.4",
-    "stylus": "^0.54.5",
+    "stylus": "^0.54.7",
     "supervisor": "^0.12.0",
     "ts-loader": "^6.0.4",
-    "tslint": "^5.18.0",
-    "typescript": "^3.5.2",
-    "webpack": "^4.35.0",
+    "tslint": "^5.19.0",
+    "typescript": "^3.6.2",
+    "webpack": "^4.39.3",
     "webpack-bundle-analyzer": "",
-    "webpack-cli": "^3.3.5",
+    "webpack-cli": "^3.3.8",
     "xgettext-js": "^1.1.0",
     "yarn": "^1.17.3"
   },
   "dependencies": {
-    "@sentry/browser": "^5.4.3",
-    "@types/blueimp-md5": "^1.1.30",
+    "@sentry/browser": "^5.6.3",
     "@types/d3": "^5.0.0",
     "@types/fixed-data-table": "^0.6.30",
     "@types/jquery": "^2.0.34",
     "@types/markdown-it": "^0.0.4",
     "@types/query-string": "^5.0.1",
-    "@types/react": "^16.8.22",
+    "@types/react": "^16.9.2",
     "@types/react-autosuggest": "^9.3.5",
-    "@types/react-dom": "^16.8.4",
+    "@types/react-dom": "^16.9.0",
     "@types/react-redux": "^4.4.35",
     "@types/react-select": "^1.2.0",
     "@types/react-virtualized": "^9.18.8",
     "@types/sanitize-html": "^1.13.31",
     "@types/socket.io-client": "^1.4.31",
     "@types/sweetalert": "^1.1.27",
-    "blueimp-md5": "^2.10.0",
+    "ts-md5": "^1.2.5",
     "d3": "^5.7.0",
     "eventemitter3": "^2.0.3",
     "jsonwebtoken": "^8.5.1",
     "markdown-it": "^8.4.2",
     "moment": "^2.24.0",
     "query-string": "^5.0.1",
-    "react": "^16.8.6",
+    "react": "^16.9.0",
     "react-autosuggest": "^9.4.2",
     "react-datetime": "^2.10.3",
-    "react-dom": "^16.8.4",
+    "react-dom": "^16.9.0",
     "react-dropzone": "^3.7.3",
     "react-number-format": "^3.0.2",
     "react-redux": "^4.4.6",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
+    "react-router": "^5.0.1",
+    "react-router-dom": "^5.0.1",
     "react-select": "^1.2.1",
     "react-virtualized": "^9.21.0",
     "redux": "^3.7.2",
     "sanitize-html": "^1.14.1",
     "slugify": "^1.3.4",
-    "webpack-bundle-analyzer": "^3.3.2",
+    "webpack-bundle-analyzer": "^3.4.1",
     "whatwg-fetch": "^3.0.0"
   }
 }
